multiplexer/requirements.txt
Flask
mysql-connector
requests
Flask-Session
Werkzeug==0.16.0